% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_wind_direction.R
\name{cal_wind_direction}
\alias{cal_wind_direction}
\title{Covert wind direction code into degrees based on the base direction.}
\usage{
cal_wind_direction(x, base_direction = c("North", "South", "East", "West"))
}
\arguments{
\item{x}{Numeric. Wind direction code to be coverted.}

\item{base_direction}{Character. Direction of 0 degree. Should be one of (E)ast, (W)est, (S)outh and (N)orth.}
}
\value{
Wind direction in degree format.
}
\description{
Covert wind direction code into degrees based on the base direction.
}
