% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\title{Construct a client to talk to the CMDC api}
\usage{
client(apikey = NULL)
}
\arguments{
\item{apikey}{A string containing the APIKEY.}
}
\value{
An S3 object with class cmdcClient that implements routines for accessing CMDC API endpoints
}
\description{
Construct a client to talk to the CMDC api
}
