% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine meteorological data into a single file.}
\usage{
combine_data(flist, outdir)
}
\arguments{
\item{flist}{Character. List of file names to be processed.}

\item{outdir}{Character. Directory to store the files.}
}
\value{
\code{NULL}
}
\description{
Combine meteorological data into a single file.
}
