% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data_}
\alias{combine_data_}
\title{Combine meteorological data into a single file.}
\usage{
combine_data_(dtype, flist, outdir, subfolder = T, maxsize = 512)
}
\arguments{
\item{dtype}{Character. Type of the meteorological variable.}

\item{flist}{Character. List of file names to be processed.}

\item{outdir}{Character. Directory to store the files.}

\item{subfolder}{Logical. \code{True} will create the subfolder according to the data type (obtained by \code{\link[cmdcr]{get_data_type}}).}

\item{maxsize}{Numeric. Maximum size of data to be read into memory. Default is 512 Mb.
Note that if the total size of files exceed \code{maxisize}. The function will run in append mode.
Please refer to \code{\link[data.table]{fwrite}} for details.}
}
\description{
Combine meteorological data into a single file.
}
