% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{fetch.cmdcClient}
\alias{fetch.cmdcClient}
\title{Fetch the currently constructed dataset}
\usage{
\method{fetch}{cmdcClient}(x)
}
\value{
A data.frame (tibble) containing requested data
}
\description{
Fetch the currently constructed dataset
}
