% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\alias{format_data.WIN}
\title{Tidy the meteorological data.}
\usage{
format_data(x, ...)

\method{format_data}{WIN}(x, base_direction = "N", outdir, prefix = "F",
  ...)
}
\arguments{
\item{x}{Charater. Full name of file to be processed. Generate with \code{\link[cmdcr]{read_fname}}.}

\item{...}{Not used yet.}

\item{base_direction}{Character. Direction of 0 degree. Should be one of (E)ast, (W)est, (S)outh and (N)orth.}

\item{outdir}{Character. Output directory.}

\item{prefix}{Character. Prefix of the output file name.}
}
\description{
Tidy the meteorological data.
}
