% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_type.R
\name{get_data_type}
\alias{get_data_type}
\title{Get the type of the meteorological variable according to the file name.}
\usage{
get_data_type(fname)
}
\arguments{
\item{fname}{Character. The full name of the file to be processed.}
}
\value{
The type of the meteorological variable.
}
\description{
Get the type of the meteorological variable according to the file name.
}
