% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_station.R
\name{get_nearest_station}
\alias{get_nearest_station}
\title{Get the nearest stations to the preferd station.}
\usage{
get_nearest_station(id, lonlat = T, dt, ...)
}
\arguments{
\item{id}{The ID of the desired station}

\item{lonlat}{logical. If TRUE, coordinates should be in degrees; 
else they should represent planar ('Euclidean') space (e.g. units of meters).
See \link[raster]{pointDistance} for details.}

\item{dt}{A data.table which contains 3 columns (i.e., ID, Lon, Lat).}

\item{...}{Other parameters which passed to \link[raster]{pointDistance}}
}
\value{
A vector with station IDs sorted by distance.
}
\description{
Get the nearest stations to the preferd station.
}
