% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cmdc_data.R
\name{import_cmdc_data}
\alias{import_cmdc_data}
\title{Import CMDC data}
\usage{
import_cmdc_data(fname)
}
\arguments{
\item{fname}{Character. Full name of the file to be processed.}
}
\value{
A \code{cmdc} typed object.
}
\description{
Import CMDC data
}
