% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{info.cmdcClient}
\alias{info.cmdcClient}
\title{Get information about the API as a whole (`info(client)`) or a specific
endpoint (`info(client, endpoint)`, for exapmle `info(client, "demographics")`)}
\usage{
info.cmdcClient(c, x = NULL)
}
\description{
Get information about the API as a whole (`info(client)`) or a specific
endpoint (`info(client, endpoint)`, for exapmle `info(client, "demographics")`)
}
