% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_type.R
\name{is_valid_fname}
\alias{is_valid_fname}
\title{Check the validity of the file name.}
\usage{
is_valid_fname(fname)
}
\arguments{
\item{fname}{Character. The full name of the file to be processed.}
}
\value{
\code{TRUE} if the file name is valid and vice versa.
}
\description{
Check the validity of the file name.
}
