% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{print.cmdcClient}
\alias{print.cmdcClient}
\title{Print information about the client and current request}
\usage{
\method{print}{cmdcClient}(x)
}
\description{
Print information about the client and current request
}
