% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_cmdc_data.R
\name{rbind_cmdc_data}
\alias{rbind_cmdc_data}
\alias{merge_cmdc_data}
\title{Combine cmdc data.table by rows or columns}
\usage{
rbind_cmdc_data(..., dtlist = NULL, dir, filepattern = c("PRS", "TEM",
  "WIN", "GST", "PRE", "RHU", "EVP", "SSD"))

merge_cmdc_data(..., dtlist = NULL, dir)
}
\arguments{
\item{...}{(optional) List of \code{data.table} objects to be combined.}

\item{dtlist}{(optional) List of \code{data.table} to be combined.}

\item{dir}{Character. The directory which contains the cmdc files.}

\item{filepattern.}{Character. Variable to be combined.}
}
\value{
A \code{data.table}.
}
\description{
Combine cmdc data.table by rows or columns

Merge cmdc data.table
}
