% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read the meteorological data.}
\usage{
read_data(fname)
}
\arguments{
\item{fname}{Character. Full name of the file to be processed.}

\item{as.data.table}{Logical. If \code{True}, a data.table will return.}
}
\description{
Read the meteorological data.
}
