% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fname.R
\name{read_fname}
\alias{read_fname}
\title{Read the file name of the meteorological data.}
\usage{
read_fname(fname)
}
\arguments{
\item{fname}{Character. Full name of the file to be processed.}
}
\description{
Read the file name of the meteorological data.
}
