% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{register.cmdcClient}
\alias{register.cmdcClient}
\title{Fetch the currently constructed dataset}
\usage{
\method{register}{cmdcClient}(x, ...)
}
\value{
A string containing the API Key
}
\description{
Fetch the currently constructed dataset
}
