% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing_data.R
\name{remove_missing_data}
\alias{remove_missing_data}
\title{Function to remove missing data in the data.table.}
\usage{
remove_missing_data(x, ...)
}
\arguments{
\item{x}{Numeric. A column of data to be processed.}

\item{...}{Not used yet.}
}
\description{
Function to remove missing data in the data.table.
}
