% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{reset.cmdcClient}
\alias{reset.cmdcClient}
\title{Reset the current request on this client}
\usage{
\method{reset}{cmdcClient}(x)
}
\value{
The client
}
\description{
Reset the current request on this client
}
