% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colnames.R
\name{set_colnames}
\alias{set_colnames}
\alias{set_colnames.PRS}
\alias{set_colnames.TEM}
\alias{set_colnames.RHU}
\alias{set_colnames.PRE}
\alias{set_colnames.EVP}
\alias{set_colnames.WIN}
\title{Set column names for each dataset.}
\usage{
set_colnames(x, ...)

\method{set_colnames}{PRS}(x, ...)

\method{set_colnames}{TEM}(x, ...)

\method{set_colnames}{RHU}(x, ...)

\method{set_colnames}{PRE}(x, ...)

\method{set_colnames}{EVP}(x, ...)

\method{set_colnames}{WIN}(x, ...)
}
\arguments{
\item{x}{A \code{cmdc} object. Created by \link[cmdcr]{import_cmdc_data}.}

\item{...}{Not used yet.}
}
\description{
Set column names for each dataset.
}
