% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\alias{format_data.WIN}
\alias{format_data.PRS}
\alias{format_data.EVP}
\alias{format_data.TEM}
\alias{format_data.RHU}
\alias{format_data.PRE}
\alias{format_data.SSD}
\alias{format_data.GST}
\title{Format the meteorological data.}
\usage{
format_data(x, ...)

\method{format_data}{WIN}(x, base_direction = "N", ...)

\method{format_data}{PRS}(x, ...)

\method{format_data}{EVP}(x, ...)

\method{format_data}{TEM}(x, ...)

\method{format_data}{RHU}(x, ...)

\method{format_data}{PRE}(x, ...)

\method{format_data}{SSD}(x, ...)

\method{format_data}{GST}(x, ...)
}
\arguments{
\item{x}{A \code{cmdc} object. Created by \link[cmdcr]{import_cmdc_data}}

\item{...}{Not used yet.}

\item{base_direction}{Character. Direction of 0 degree. Should be one of (E)ast, (W)est, (S)outh and (N)orth.}
}
\description{
Format the meteorological data.
}
